health: Move endpoint IP to node package
[ upstream commit ed934cb958c52515be37998038301229959a0fd2 ]

The Cilium health endpoint IP is the only static local node field which
is not accessible from the node package. To avoid code modifying the
`NodeDiscovery.LocalNode` field (which will become more dynamic in a
subsequent mode), the health IP is now moved to the node package with a
corresponding getter and setter like all other static node fields.

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>
Signed-off-by: Gilberto Bertin <jibi@cilium.io>