nodediscovery: Unify source of truth for node name
[ upstream commit 9c1cb4856322a80174799dae142ef761e51e2310 ]

Clustermesh related code is relying on
`NodeDiscovery.LocalNode.FullName()` to obtain the globally unique name
of the local node. This field is initialized from `nodeTypes.GetName()`.
This code removes all uses of `LocalNode.FullName()` and replaces it
with an equivalent call to `nodeTypes.GetAbsoluteNodeName()`. This
allows us to make the `NodeDiscovery.LocalNode` private in a subsequent
commit.

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>
Signed-off-by: Gilberto Bertin <jibi@cilium.io>