daemon: Deprecate KPR=probe
The option is a never ending source of confusion. Also, after changing
system configuration, the "probe" option could unexpectedly enable some
KPR features which might be not expected by users.

Deprecate the "probe" option for now. In v1.13 we will make the option
to switch to "disabled". Unfortunately, we cannot just remove the
option, as it was used as a default in the Helm charts.

Signed-off-by: Martynas Pumputis <m@lambda.lt>